Restructure project for Dockerization
Adds:
- Dockerfile
- Pipfile
- Pipfile.lock
- pyproject.toml
- requirements.txt

Requirements.txt is generated by running docker build.
Dockerfile generates a package of all the dependencies. Currently, the main code is packaged with
a simple zip, and manually uploaded to lambda. The two zips are combined in the lambda function
as layers.
