+++ Pipfile
@@ -1,0 +1,15 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+opencv-python = "*"
+numpy = "*"
+picfix-core = {editable = true, path = "."}
+flask = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.7"
